Rename IndexPage to Home and reorganize files
