+++ client/src/containers/Home.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 import MapContainer from './MapContainer';
-import LocationList from '../components/LocationList';
+import LocationList from '../containers/LocationList';
 import NewLocationForm from '../components/NewLocationForm';
 import { connect } from 'react-redux';
 import { createLocation, deleteLocation } from '../actions/thunks';
 
-export class IndexPage extends Component {
+export class Home extends Component {
 
   constructor(props) {
     super(props);
@@ -42,4 +42,4 @@ function mapDispatchToProps(dispatch) {
           }
 }
 
-export default connect(null, mapDispatchToProps)(IndexPage)
+export default connect(null, mapDispatchToProps)(Home)
